refactor: Handle all image and fallback loading in the same component
