+++ src/drive/containers/File.jsx
@@ -10,7 +10,7 @@ import { translate } from 'cozy-ui/react/I18n'
 import RenameInput from '../ducks/files/RenameInput'
 import { isDirectory } from '../ducks/files/files'
 import Spinner from 'cozy-ui/react/Spinner'
-import Preview from '../components/Preview'
+import { ImageLoader } from 'components/Image'
 import { Button, Icon, withBreakpoints, MidEllipsis } from 'cozy-ui/react'
 import { SharedBadge, SharedStatus } from 'sharing'
 import { getFileTypeFromMime } from 'drive/lib/getFileTypeFromMime'
@@ -109,7 +109,16 @@ const FileName = ({
   const url = cozy.client._url
   return (
     <div className={classes}>
-      <Preview file={attributes} size="small" />
+      <ImageLoader
+        file={attributes}
+        size="small"
+        render={src => (
+          <div
+            className={styles['fil-file-preview']}
+            style={`background-image: url(${src});`}
+          />
+        )}
+      />
       <SharedBadge
         docId={attributes.id}
         className={styles['fil-content-shared']}
